Changed properties int Grade to double Grade, added migration

+++ PRIS.Core.Library/Entities/ConversationResult.cs
@@ -4,7 +4,7 @@ namespace PRIS.Core.Library.Entities
 {
     public class ConversationResult : EntityBase
     {
-        public int? Grade { get; set; }
+        public double? Grade { get; set; }
         public string Comment { get; set; }
 
         public List<Student> Students { get; set; }
